+++ src/Configuration.ts
@@ -1,7 +1,7 @@
 // configuration lives in ../config.json
 // this is the type definition for it, with example values in comments
 
-export interface ConfigType {
+export interface Configuration {
     mqttServer: string // "mqtt://127.0.0.1"
     mqttOptions: any // null
     mqttTopic: string // "laser/maintenance"
@@ -12,4 +12,4 @@ export interface ConfigType {
     }
 }
 
-export const config: ConfigType = require('../config.json');
+export const config: Configuration = require('../config.json');
