+++ test/testutil.js
@@ -1,14 +1,14 @@
 // @ts-nocheck
 'use strict'
 
-const { LaserAccessApp } = require('../dist/app')
+const { LaserWebApp } = require('../dist/LaserWebApp');
 
 let mainApp;
 let init = false;
 
 const getApp = function () {
   if (!mainApp) {
-    mainApp = new LaserAccessApp();
+    mainApp = new LaserWebApp();
 
     mainApp.expressApp.use('/mock500', function (_req, _res, next) {
       next('Unittest error')
